ECP: Respecify event loop to account for duplicate states and redundancies
This changes the event specification by adding a single additional event CSWP while allowing for the stimuli inputs to be more abstract, taking advantage of duplicate state in the specification and the global application state of `ac`, `fn`, and `tie` in particular.

There was significant redundancy in the existing specification due to encoding duplicate state information already encoded in the application state. This respecification rectifies this.

Additionally, this also makes the underlying output slightly more immediate mode, by permitting some responses to act as no-ops when appropriate.